api/requirements.txt
# azure-core==1.33.0
# azure-cosmos==4.9.0
# azure-functions==1.23.0
# azure-identity==1.5.0
# azure-keyvault-secrets==4.9.0
# azure-storage-blob==12.25.1
# certifi==2025.4.26
# cffi==1.17.1
# charset-normalizer==3.4.2
# cryptography==45.0.4
# ecdsa==0.19.1
# idna==3.10
# isodate==0.7.2
# MarkupSafe==3.0.2
# msal==1.32.3
# msal-extensions==0.3.1
# pycparser==2.22
# PyJWT==2.10.1
# python-http-client==3.3.7
# requests==2.32.4
# sendgrid==6.12.3
# six==1.17.0
# typing_extensions==4.14.0
# urllib3==2.4.0
# Werkzeug==3.1.3

azure-core==1.33.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.5.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
ecdsa==0.19.1
idna==3.10
isodate==0.7.2
MarkupSafe>=2.1.1
msal==1.32.3
msal-extensions==0.3.1
pycparser==2.22
PyJWT>=2.0.0,<3.0.0
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.3
six==1.17.0
typing_extensions==4.13.2
urllib3==1.26.15
Werkzeug>=2.3.7,<=3.1.3
sendgrid==6.11.0
